Add SRIOV CI tests for downwardAPI and VF flags
Add some test code to test the downwardAPI, which mounts:
/etc/podnetinfo/labels
/etc/podnetinfo/annotations
in the Pod.

Also, add some test code to test following VF flags:
spoofchk
max_tx_rate
trust.

Verify that above flags are reset to default when the VF
is released from the Pod.

Note: This test code is yet to be executed on env with
SRIOV hardware. May require a tweak or two....

Signed-off-by: vpickard <vpickard@redhat.com>