e2e-sriov: Allow more time for VFs to be allocatable
The last SR-IOV job failed because it timed out after
waiting 20 seconds for VFs to be allocatable. System
seemed really busy/loaded yesterday, suspect that load
was higher than usual.

Increase the timer from 20s to 60s to see if this is
enough time for VFs to become allocatable when system
is loaded.

Signed-off-by: vpickard <vpickard@redhat.com>